feat : create LottoTicket and separate Lotto's functions
